Further consolidate scenario-detail API
