+++ ui/api/create_api.py
@@ -2,8 +2,7 @@
 
 # RESTful API resources
 from ui.api.resources.home import ServerStatus
-from ui.api.resources.scenario_detail import ScenarioDetailName, \
-  ScenarioDetailAll, ScenarioDetailAPI
+from ui.api.resources.scenario_detail import ScenarioDetailAPI
 from ui.api.resources.scenario_results import \
   ScenarioResultsProjectCapacity, ScenarioResultsProjectRetirements, \
   ScenarioResultsProjectNewBuild, ScenarioResultsProjectDispatch, \
@@ -77,19 +76,6 @@ def add_scenario_detail_resources(api, db_path):
     api.add_resource(
         ScenarioDetailAPI,
         '/scenarios/<scenario_id>',
-        resource_class_kwargs={'db_path': db_path}
-    )
-
-    # All
-    api.add_resource(
-        ScenarioDetailAll,
-        '/scenarios/<scenario_id>/all',
-        resource_class_kwargs={'db_path': db_path}
-    )
-    # Name
-    api.add_resource(
-        ScenarioDetailName,
-        '/scenarios/<scenario_id>/name',
         resource_class_kwargs={'db_path': db_path}
     )
