+++ ui/api/resources/scenario_detail.py
@@ -16,16 +16,41 @@ class ScenarioDetailAPI(Resource):
 
     def get(self, scenario_id):
         io, c = connect_to_database(db_path=self.db_path)
+
+        scenario_detail_api = dict()
+
+        # Get the scenario name
+        scenario_name = c.execute(
+          "SELECT scenario_name "
+          "FROM scenarios "
+          "WHERE scenario_id = {}".format(scenario_id)
+        ).fetchone()[0]
+
+        scenario_detail_api["scenarioName"] = scenario_name
+
+        # Get the values for scenario-edit
+        scenario_edit_query = c.execute(
+          "SELECT * "
+          "FROM scenarios_view "
+          "WHERE scenario_id = {}".format(scenario_id)
+        )
+
+        column_names = [s[0] for s in scenario_edit_query.description]
+        column_values = list(list(scenario_edit_query)[0])
+        scenario_edit_api = dict(zip(column_names, column_values))
+
+        scenario_detail_api["editScenarioValues"] = scenario_edit_api
+
         all_tables = c.execute(
             """SELECT ui_table 
             FROM ui_scenario_detail_table_metadata
             ORDER BY ui_table_id ASC;"""
         ).fetchall()
 
-        scenario_detail_api = list()
+        scenario_detail_api["scenarioDetailTables"] = list()
 
         for ui_table in all_tables:
-            scenario_detail_api.append(
+            scenario_detail_api["scenarioDetailTables"].append(
                 get_scenario_detail(
                   cursor=c,
                   scenario_id=scenario_id,
@@ -34,47 +59,6 @@ class ScenarioDetailAPI(Resource):
             )
 
         return scenario_detail_api
-
-
-class ScenarioDetailAll(Resource):
-    """
-    All selections for a scenario.
-    """
-
-    def __init__(self, **kwargs):
-        self.db_path = kwargs["db_path"]
-
-    def get(self, scenario_id):
-        io, c = connect_to_database(db_path=self.db_path)
-        scenario_detail_query = c.execute(
-            "SELECT * "
-            "FROM scenarios_view "
-            "WHERE scenario_id = {}".format(scenario_id)
-        )
-
-        column_names = [s[0] for s in scenario_detail_query.description]
-        column_values = list(list(scenario_detail_query)[0])
-        scenario_detail_api = dict(zip(column_names, column_values))
-
-        return scenario_detail_api
-
-
-class ScenarioDetailName(Resource):
-    """
-    The name of the a scenario by scenario ID
-    """
-    def __init__(self, **kwargs):
-        self.db_path = kwargs["db_path"]
-
-    def get(self, scenario_id):
-        io, c = connect_to_database(db_path=self.db_path)
-        scenario_name = c.execute(
-            "SELECT scenario_name "
-            "FROM scenarios "
-            "WHERE scenario_id = {}".format(scenario_id)
-        ).fetchone()[0]
-
-        return scenario_name
 
 
 def get_scenario_detail(cursor, scenario_id, ui_table_name_in_db):
@@ -95,12 +79,12 @@ def get_scenario_detail(cursor, scenario_id, ui_table_name_in_db):
         WHERE ui_table = '{}';""".format(ui_table_name_in_db)
     ).fetchone()
 
-    scenario_detail_api = dict()
-    scenario_detail_api["uiTableNameInDB"] = table_caption[0]
-    scenario_detail_api["scenarioDetailTableCaption"] = table_caption[1]
+    scenario_detail_table_api = dict()
+    scenario_detail_table_api["uiTableNameInDB"] = table_caption[0]
+    scenario_detail_table_api["scenarioDetailTableCaption"] = table_caption[1]
 
     # Get the metadata and value for the rows
-    scenario_detail_api["scenarioDetailTableRows"] = list()
+    scenario_detail_table_api["scenarioDetailTableRows"] = list()
 
     row_metadata = c.execute(
         """SELECT ui_table_row, ui_row_caption, ui_row_db_scenarios_view_column, 
@@ -116,11 +100,11 @@ def get_scenario_detail(cursor, scenario_id, ui_table_name_in_db):
               WHERE scenario_id = {}""".format(row[2], scenario_id)
         ).fetchone()[0]
 
-        scenario_detail_api["scenarioDetailTableRows"].append({
+        scenario_detail_table_api["scenarioDetailTableRows"].append({
             'uiRowNameInDB': row[0],
             'rowCaption': row[1],
             'rowValue': row_value,
             'inputTable':  row[3]
         })
 
-    return scenario_detail_api
+    return scenario_detail_table_api
