+++ ui/src/app/scenario-detail/scenario-detail.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { ScenarioDetail } from './scenario-detail';
+import { ScenarioDetailAPI } from './scenario-detail';
 
 @Injectable({
   providedIn: 'root'
@@ -14,14 +14,8 @@ export class ScenarioDetailService {
 
   private scenariosBaseURL = 'http://127.0.0.1:8080/scenarios/';
 
-  getScenarioName(scenarioID: number): Observable<string> {
-    return this.http.get<string>(
-      `${this.scenariosBaseURL}${scenarioID}/name`
-    );
-  }
-
-  getScenarioDetailAPI(scenarioID: number): Observable<ScenarioDetail[]> {
-    return this.http.get<ScenarioDetail[]>(
+  getScenarioDetailAPI(scenarioID: number): Observable<ScenarioDetailAPI> {
+    return this.http.get<ScenarioDetailAPI>(
       `${this.scenariosBaseURL}${scenarioID}`
     );
   }
